+++ Source/Scene/Model.js
@@ -20,6 +20,7 @@ define([
         '../Core/getMagic',
         '../Core/getStringFromTypedArray',
         '../Core/IndexDatatype',
+        '../Core/isArray',
         '../Core/loadCRN',
         '../Core/loadImageFromTypedArray',
         '../Core/loadKTX',
@@ -70,6 +71,7 @@ define([
         './ModelMesh',
         './ModelNode',
         './ModelUtility',
+        './OctahedralProjectedCubeMap',
         './processModelMaterialsCommon',
         './processPbrMaterials',
         './SceneMode',
@@ -96,6 +98,7 @@ define([
         getMagic,
         getStringFromTypedArray,
         IndexDatatype,
+        isArray,
         loadCRN,
         loadImageFromTypedArray,
         loadKTX,
@@ -146,6 +149,7 @@ define([
         ModelMesh,
         ModelNode,
         ModelUtility,
+        OctahedralProjectedCubeMap,
         processModelMaterialsCommon,
         processPbrMaterials,
         SceneMode,
@@ -288,6 +292,9 @@ define([
      * @param {Boolean} [options.dequantizeInShader=true] Determines if a {@link https://github.com/google/draco|Draco} encoded model is dequantized on the GPU. This decreases total memory usage for encoded models.
      * @param {Cartesian2} [options.imageBasedLightingFactor=Cartesian2(1.0, 1.0)] Scales diffuse and specular image-based lighting from the earth, sky, atmosphere and star skybox.
      * @param {Cartesian3} [options.lightColor] The color and intensity of the sunlight used to shade the model.
+     * @param {Number} [options.luminanceAtZenith=0.5] The sun's luminance at the zenith in kilo candela per meter squared to use for this model's procedural environment map.
+     * @param {Cartesian3[]} [options.sphericalHarmonicCoefficients] The third order spherical harmonic coefficients used for the diffuse color of image-based lighting.
+     * @param {String} [options.specularEnvironmentMaps] A URL to a KTX file that contains a cube map of the specular lighting and the convoluted specular mipmaps.
      *
      * @see Model.fromGltf
      *
@@ -668,7 +675,19 @@ define([
         this._imageBasedLightingFactor = new Cartesian2(1.0, 1.0);
         Cartesian2.clone(options.imageBasedLightingFactor, this._imageBasedLightingFactor);
         this._lightColor = Cartesian3.clone(options.lightColor);
-        this._regenerateShaders = false;
+
+        this._luminanceAtZenith = undefined;
+        this.luminanceAtZenith = defaultValue(options.luminanceAtZenith, 0.5);
+
+        this._sphericalHarmonicCoefficients = options.sphericalHarmonicCoefficients;
+        this._specularEnvironmentMaps = options.specularEnvironmentMaps;
+        this._shouldUpdateSpecularMapAtlas = true;
+        this._specularEnvironmentMapAtlas = undefined;
+
+        this._useDefaultSphericalHarmonics = false;
+        this._useDefaultSpecularMaps = false;
+
+        this._shouldRegenerateShaders = false;
     }
 
     defineProperties(Model.prototype, {
@@ -833,7 +852,7 @@ define([
          * // Play all animations at half-speed when the model is ready to render
          * Cesium.when(model.readyPromise).then(function(model) {
          *   model.activeAnimations.addAll({
-         *     speedup : 0.5
+         *     multiplier : 0.5
          *   });
          * }).otherwise(function(error){
          *   window.alert(error);
@@ -1109,8 +1128,8 @@ define([
                 Check.typeOf.number.greaterThanOrEquals('imageBasedLightingFactor.y', value.y, 0.0);
                 Check.typeOf.number.lessThanOrEquals('imageBasedLightingFactor.y', value.y, 1.0);
                 //>>includeEnd('debug');
-                this._regenerateShaders = this._regenerateShaders || (this._imageBasedLightingFactor.x > 0.0 && value.x === 0.0) || (this._imageBasedLightingFactor.x === 0.0 && value.x > 0.0);
-                this._regenerateShaders = this._regenerateShaders || (this._imageBasedLightingFactor.y > 0.0 && value.y === 0.0) || (this._imageBasedLightingFactor.y === 0.0 && value.y > 0.0);
+                this._shouldRegenerateShaders = this._shouldRegenerateShaders || (this._imageBasedLightingFactor.x > 0.0 && value.x === 0.0) || (this._imageBasedLightingFactor.x === 0.0 && value.x > 0.0);
+                this._shouldRegenerateShaders = this._shouldRegenerateShaders || (this._imageBasedLightingFactor.y > 0.0 && value.y === 0.0) || (this._imageBasedLightingFactor.y === 0.0 && value.y > 0.0);
                 Cartesian2.clone(value, this._imageBasedLightingFactor);
             }
         },
@@ -1136,8 +1155,83 @@ define([
                 if (value === lightColor || Cartesian3.equals(value, lightColor)) {
                     return;
                 }
-                this._regenerateShaders = this._regenerateShaders || (defined(lightColor) && !defined(value)) || (defined(value) && !defined(lightColor));
+                this._shouldRegenerateShaders = this._shouldRegenerateShaders || (defined(lightColor) && !defined(value)) || (defined(value) && !defined(lightColor));
                 this._lightColor = Cartesian3.clone(value, lightColor);
+            }
+        },
+
+        /**
+         * The sun's luminance at the zenith in kilo candela per meter squared to use for this model's procedural environment map.
+         * This is used when {@link Model#specularEnvironmentMaps} and {@link Model#sphericalHarmonicCoefficients} are not defined.
+         *
+         * @memberof Model.prototype
+         *
+         * @demo {@link https://cesiumjs.org/Cesium/Apps/Sandcastle/index.html?src=Image-Based Lighting.html|Sandcastle Image Based Lighting Demo}
+         * @type {Number}
+         * @default 0.5
+         */
+        luminanceAtZenith : {
+            get : function() {
+                return this._luminanceAtZenith;
+            },
+            set : function(value) {
+                var lum = this._luminanceAtZenith;
+                if (value === lum) {
+                    return;
+                }
+                this._shouldRegenerateShaders = this._shouldRegenerateShaders || (defined(lum) && !defined(value)) || (defined(value) && !defined(lum));
+                this._luminanceAtZenith = value;
+            }
+        },
+
+        /**
+         * The third order spherical harmonic coefficients used for the diffuse color of image-based lighting. When <code>undefined</code>, a diffuse irradiance
+         * computed from the atmosphere color is used.
+         * <p>
+         * There are nine <code>Cartesian3</code> coefficients.
+         * The order of the coefficients is: L<sub>00</sub>, L<sub>1-1</sub>, L<sub>10</sub>, L<sub>11</sub>, L<sub>2-2</sub>, L<sub>2-1</sub>, L<sub>20</sub>, L<sub>21</sub>, L<sub>22</sub>
+         * </p>
+         *
+         * These values can be obtained by preprocessing the environment map using the <code>cmgen</code> tool of
+         * {@link https://github.com/google/filament/releases | Google's Filament project}. This will also generate a KTX file that can be
+         * supplied to {@link Model#specularEnvironmentMaps}.
+         *
+         * @memberof Model.prototype
+         *
+         * @type {Cartesian3[]}
+         * @demo {@link https://cesiumjs.org/Cesium/Apps/Sandcastle/index.html?src=Image-Based Lighting.html|Sandcastle Image Based Lighting Demo}
+         * @see {@link https://graphics.stanford.edu/papers/envmap/envmap.pdf|An Efficient Representation for Irradiance Environment Maps}
+         */
+        sphericalHarmonicCoefficients : {
+            get : function() {
+                return this._sphericalHarmonicCoefficients;
+            },
+            set : function(value) {
+                //>>includeStart('debug', pragmas.debug);
+                if (defined(value) && (!isArray(value) || value.length !== 9)) {
+                    throw new DeveloperError('sphericalHarmonicCoefficients must be an array of 9 Cartesian3 values.');
+                }
+                //>>includeEnd('debug');
+                this._sphericalHarmonicCoefficients = value;
+                this._shouldRegenerateShaders = true;
+            }
+        },
+
+        /**
+         * A URL to a KTX file that contains a cube map of the specular lighting and the convoluted specular mipmaps.
+         *
+         * @memberof Model.prototype
+         * @demo {@link https://cesiumjs.org/Cesium/Apps/Sandcastle/index.html?src=Image-Based Lighting.html|Sandcastle Image Based Lighting Demo}
+         * @type {String}
+         * @see Model#sphericalHarmonicCoefficients
+         */
+        specularEnvironmentMaps : {
+            get : function() {
+                return this._specularEnvironmentMaps;
+            },
+            set : function(value) {
+                this._shouldUpdateSpecularMapAtlas = value !== this._specularEnvironmentMaps;
+                this._specularEnvironmentMaps = value;
             }
         }
     });
@@ -2002,7 +2096,8 @@ define([
             drawFS = 'uniform vec4 czm_pickColor;\n' + drawFS;
         }
 
-        if (model._imageBasedLightingFactor.x > 0.0 || model._imageBasedLightingFactor.y > 0.0) {
+        var useIBL = model._imageBasedLightingFactor.x > 0.0 || model._imageBasedLightingFactor.y > 0.0;
+        if (useIBL) {
             drawFS = '#define USE_IBL_LIGHTING \n\n' + drawFS;
         }
 
@@ -2019,6 +2114,29 @@ define([
                 '    non_gamma_corrected_main(); \n' +
                 '    gl_FragColor = czm_gammaCorrect(gl_FragColor); \n' +
                 '} \n';
+        }
+
+        var usesSH = defined(model._sphericalHarmonicCoefficients) || model._useDefaultSphericalHarmonics;
+        var usesSM = (defined(model._specularEnvironmentMapAtlas) && model._specularEnvironmentMapAtlas.ready) || model._useDefaultSpecularMaps;
+        var addMatrix = usesSH || usesSM || useIBL;
+        if (addMatrix) {
+            drawFS = 'uniform mat4 gltf_clippingPlanesMatrix; \n' + drawFS;
+        }
+
+        if (defined(model._sphericalHarmonicCoefficients)) {
+            drawFS = '#define DIFFUSE_IBL \n' + '#define CUSTOM_SPHERICAL_HARMONICS \n' + 'uniform vec3 gltf_sphericalHarmonicCoefficients[9]; \n' + drawFS;
+        } else if (model._useDefaultSphericalHarmonics) {
+            drawFS = '#define DIFFUSE_IBL \n' + drawFS;
+        }
+
+        if (defined(model._specularEnvironmentMapAtlas) && model._specularEnvironmentMapAtlas.ready) {
+            drawFS = '#define SPECULAR_IBL \n' + '#define CUSTOM_SPECULAR_IBL \n' + 'uniform sampler2D gltf_specularMap; \n' + 'uniform vec2 gltf_specularMapSize; \n' + 'uniform float gltf_maxSpecularLOD; \n' + drawFS;
+        } else if (model._useDefaultSpecularMaps) {
+            drawFS = '#define SPECULAR_IBL \n' + drawFS;
+        }
+
+        if (defined(model._luminanceAtZenith)) {
+            drawFS = '#define USE_SUN_LUMINANCE \n' + 'uniform float gltf_luminanceAtZenith;\n' + drawFS;
         }
 
         createAttributesAndProgram(programId, techniqueId, drawFS, drawVS, model, context);
@@ -2063,7 +2181,8 @@ define([
             drawFS = 'uniform vec4 czm_pickColor;\n' + drawFS;
         }
 
-        if (model._imageBasedLightingFactor.x > 0.0 || model._imageBasedLightingFactor.y > 0.0) {
+        var useIBL = model._imageBasedLightingFactor.x > 0.0 || model._imageBasedLightingFactor.y > 0.0;
+        if (useIBL) {
             drawFS = '#define USE_IBL_LIGHTING \n\n' + drawFS;
         }
 
@@ -2080,6 +2199,29 @@ define([
                 '    non_gamma_corrected_main(); \n' +
                 '    gl_FragColor = czm_gammaCorrect(gl_FragColor); \n' +
                 '} \n';
+        }
+
+        var usesSH = defined(model._sphericalHarmonicCoefficients) || model._useDefaultSphericalHarmonics;
+        var usesSM = (defined(model._specularEnvironmentMapAtlas) && model._specularEnvironmentMapAtlas.ready) || model._useDefaultSpecularMaps;
+        var addMatrix = !addClippingPlaneCode && (usesSH || usesSM || useIBL);
+        if (addMatrix) {
+            drawFS = 'uniform mat4 gltf_clippingPlanesMatrix; \n' + drawFS;
+        }
+
+        if (defined(model._sphericalHarmonicCoefficients)) {
+            drawFS = '#define DIFFUSE_IBL \n' + '#define CUSTOM_SPHERICAL_HARMONICS \n' + 'uniform vec3 gltf_sphericalHarmonicCoefficients[9]; \n' + drawFS;
+        } else if (model._useDefaultSphericalHarmonics) {
+            drawFS = '#define DIFFUSE_IBL \n' + drawFS;
+        }
+
+        if (defined(model._specularEnvironmentMapAtlas) && model._specularEnvironmentMapAtlas.ready) {
+            drawFS = '#define SPECULAR_IBL \n' + '#define CUSTOM_SPECULAR_IBL \n' + 'uniform sampler2D gltf_specularMap; \n' + 'uniform vec2 gltf_specularMapSize; \n' + 'uniform float gltf_maxSpecularLOD; \n' + drawFS;
+        } else if (model._useDefaultSpecularMaps) {
+            drawFS = '#define SPECULAR_IBL \n' + drawFS;
+        }
+
+        if (defined(model._luminanceAtZenith)) {
+            drawFS = '#define USE_SUN_LUMINANCE \n' + 'uniform float gltf_luminanceAtZenith;\n' + drawFS;
         }
 
         createAttributesAndProgram(programId, techniqueId, drawFS, drawVS, model, context);
@@ -2917,10 +3059,11 @@ define([
     function createClippingPlanesMatrixFunction(model) {
         return function() {
             var clippingPlanes = model.clippingPlanes;
-            if (!defined(clippingPlanes)) {
+            if (!defined(clippingPlanes) && !defined(model._sphericalHarmonicCoefficients) && !defined(model._specularEnvironmentMaps)) {
                 return Matrix4.IDENTITY;
             }
-            return Matrix4.multiply(model._clippingPlaneModelViewMatrix, clippingPlanes.modelMatrix, scratchClippingPlaneMatrix);
+            var modelMatrix = defined(clippingPlanes) ? clippingPlanes.modelMatrix : Matrix4.IDENTITY;
+            return Matrix4.multiply(model._clippingPlaneModelViewMatrix, modelMatrix, scratchClippingPlaneMatrix);
         };
     }
 
@@ -2959,6 +3102,36 @@ define([
     function createLightColorFunction(model) {
         return function() {
             return model._lightColor;
+        };
+    }
+
+    function createLuminanceAtZenithFunction(model) {
+        return function() {
+            return model.luminanceAtZenith;
+        };
+    }
+
+    function createSphericalHarmonicCoefficientsFunction(model) {
+        return function() {
+            return model._sphericalHarmonicCoefficients;
+        };
+    }
+
+    function createSpecularEnvironmentMapFunction(model) {
+        return function() {
+            return model._specularEnvironmentMapAtlas.texture;
+        };
+    }
+
+    function createSpecularEnvironmentMapSizeFunction(model) {
+        return function() {
+            return model._specularEnvironmentMapAtlas.texture.dimensions;
+        };
+    }
+
+    function createSpecularEnvironmentMapLOD(model) {
+        return function() {
+            return model._specularEnvironmentMapAtlas.maximumMipmapLevel;
         };
     }
 
@@ -3052,11 +3225,16 @@ define([
             uniformMap = combine(uniformMap, {
                 gltf_color : createColorFunction(model),
                 gltf_colorBlend : createColorBlendFunction(model),
-                gltf_clippingPlanes: createClippingPlanesFunction(model),
-                gltf_clippingPlanesEdgeStyle: createClippingPlanesEdgeStyleFunction(model),
-                gltf_clippingPlanesMatrix: createClippingPlanesMatrixFunction(model),
+                gltf_clippingPlanes : createClippingPlanesFunction(model),
+                gltf_clippingPlanesEdgeStyle : createClippingPlanesEdgeStyleFunction(model),
+                gltf_clippingPlanesMatrix : createClippingPlanesMatrixFunction(model),
                 gltf_iblFactor : createIBLFactorFunction(model),
-                gltf_lightColor : createLightColorFunction(model)
+                gltf_lightColor : createLightColorFunction(model),
+                gltf_sphericalHarmonicCoefficients : createSphericalHarmonicCoefficientsFunction(model),
+                gltf_specularMap : createSpecularEnvironmentMapFunction(model),
+                gltf_specularMapSize : createSpecularEnvironmentMapSizeFunction(model),
+                gltf_maxSpecularLOD : createSpecularEnvironmentMapLOD(model),
+                gltf_luminanceAtZenith : createLuminanceAtZenithFunction(model)
             });
 
             // Allow callback to modify the uniformMap
@@ -4299,6 +4477,37 @@ define([
             }
         }
 
+        if (this._shouldUpdateSpecularMapAtlas) {
+            this._shouldUpdateSpecularMapAtlas = false;
+            this._specularEnvironmentMapAtlas = this._specularEnvironmentMapAtlas && this._specularEnvironmentMapAtlas.destroy();
+            this._specularEnvironmentMapAtlas = undefined;
+            if (defined(this._specularEnvironmentMaps)) {
+                this._specularEnvironmentMapAtlas = new OctahedralProjectedCubeMap(this._specularEnvironmentMaps);
+                var that = this;
+                this._specularEnvironmentMapAtlas.readyPromise.then(function() {
+                    that._shouldRegenerateShaders = true;
+                });
+            }
+
+            // Regenerate shaders to not use an environment map. Will be set to true again if there was a new environment map and it is ready.
+            this._shouldRegenerateShaders = true;
+        }
+
+        if (defined(this._specularEnvironmentMapAtlas)) {
+            this._specularEnvironmentMapAtlas.update(frameState);
+        }
+
+        var recompileWithDefaultAtlas = !defined(this._specularEnvironmentMapAtlas) && defined(frameState.specularEnvironmentMaps) && !this._useDefaultSpecularMaps;
+        var recompileWithoutDefaultAtlas = !defined(frameState.specularEnvironmentMaps) && this._useDefaultSpecularMaps;
+
+        var recompileWithDefaultSHCoeffs = !defined(this._sphericalHarmonicCoefficients) && defined(frameState.sphericalHarmonicCoefficients) && !this._useDefaultSphericalHarmonics;
+        var recompileWithoutDefaultSHCoeffs = !defined(frameState.sphericalHarmonicCoefficients) && this._useDefaultSphericalHarmonics;
+
+        this._shouldRegenerateShaders = this._shouldRegenerateShaders || recompileWithDefaultAtlas || recompileWithoutDefaultAtlas || recompileWithDefaultSHCoeffs || recompileWithoutDefaultSHCoeffs;
+
+        this._useDefaultSpecularMaps = !defined(this._specularEnvironmentMapAtlas) && defined(frameState.specularEnvironmentMaps);
+        this._useDefaultSphericalHarmonics = !defined(this._sphericalHarmonicCoefficients) && defined(frameState.sphericalHarmonicCoefficients);
+
         var silhouette = hasSilhouette(this, frameState);
         var translucent = isTranslucent(this);
         var invisible = isInvisible(this);
@@ -4375,13 +4584,20 @@ define([
             // Regenerate shaders if ClippingPlaneCollection state changed or it was removed
             var clippingPlanes = this._clippingPlanes;
             var currentClippingPlanesState = 0;
-            if (defined(clippingPlanes) && clippingPlanes.enabled && clippingPlanes.length > 0) {
+            var useClippingPlanes = defined(clippingPlanes) && clippingPlanes.enabled && clippingPlanes.length > 0;
+            var usesSH = defined(this._sphericalHarmonicCoefficients) || this._useDefaultSphericalHarmonics;
+            var usesSM = (defined(this._specularEnvironmentMapAtlas) && this._specularEnvironmentMapAtlas.ready) || this._useDefaultSpecularMaps;
+            if (useClippingPlanes || usesSH || usesSM) {
                 var clippingPlanesOriginMatrix = defaultValue(this.clippingPlanesOriginMatrix, modelMatrix);
                 Matrix4.multiply(context.uniformState.view3D, clippingPlanesOriginMatrix, this._clippingPlaneModelViewMatrix);
+            }
+
+            if (useClippingPlanes) {
                 currentClippingPlanesState = clippingPlanes.clippingPlanesState;
             }
 
-            var shouldRegenerateShaders = this._clippingPlanesState !== currentClippingPlanesState || this._regenerateShaders;
+            var shouldRegenerateShaders = this._shouldRegenerateShaders;
+            shouldRegenerateShaders = shouldRegenerateShaders || this._clippingPlanesState !== currentClippingPlanesState;
             this._clippingPlanesState = currentClippingPlanesState;
 
             // Regenerate shaders if color shading changed from last update
@@ -4397,8 +4613,6 @@ define([
                 updateColor(this, frameState, false);
                 updateSilhouette(this, frameState, false);
             }
-
-            this._regenerateShaders = false;
         }
 
         if (justLoaded) {
@@ -4492,7 +4706,9 @@ define([
         destroyIfNotCached(rendererResources, cachedRendererResources);
 
         var programId;
-        if (isClippingEnabled(model) || isColorShadingEnabled(model) || model._regenerateShaders) {
+        if (isClippingEnabled(model) || isColorShadingEnabled(model) || model._shouldRegenerateShaders) {
+            model._shouldRegenerateShaders = false;
+
             rendererResources.programs = {};
             rendererResources.silhouettePrograms = {};
 
@@ -4610,6 +4826,8 @@ define([
             clippingPlaneCollection.destroy();
         }
         this._clippingPlanes = undefined;
+
+        this._specularEnvironmentMapAtlas = this._specularEnvironmentMapAtlas && this._specularEnvironmentMapAtlas.destroy();
 
         return destroyObject(this);
     };